backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.5
sqlalchemy==2.0.41
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.16.1
python-dotenv==1.1.0
requests==2.32.3
httpx==0.28.1
redis==6.2.0
pytest==8.3.5
pytest-asyncio==1.0.0
gunicorn==23.0.0
loguru==0.7.3
tenacity==9.1.2
google-cloud-pubsub==2.29.0
pydantic-settings==2.9.1
python-decouple==3.8

# Gemini Google API
google-generativeai==0.8.5

# LangChain e scraping
langchain==0.3.25
langchain-community==0.3.24
langchain-google-genai==2.0.10
beautifulsoup4==4.13.4
lxml==5.4.0

# Outros úteis (comuns em FastAPI moderno)
python-multipart==0.0.20
email-validator==2.2.0
jinja2==3.1.6
